test: move test file to the right place
