+++ test/app/extend/request.test.js
@@ -5,10 +5,10 @@ const should = require('should');
 const merge = require('merge-descriptors');
 const urllib = require('urllib');
 const request = require('supertest');
-const utils = require('../../../../utils');
-const requestExt = require('../../../../../app/extend/request');
+const utils = require('../../utils');
+const requestExt = require('../../../app/extend/request');
 
-describe('test/lib/core/app/extend/request.test.js', () => {
+describe('test/app/extend/request.test.js', () => {
   afterEach(mm.restore);
 
   describe('req.host', () => {
