+++ test/app/middleware/body_parser.test.js
@@ -3,9 +3,9 @@
 const querystring = require('querystring');
 const should = require('should');
 const request = require('supertest');
-const utils = require('../../../../utils');
+const utils = require('../../utils');
 
-describe('test/lib/core/app/middleware/body_parser.test.js', () => {
+describe('test/app/middleware/body_parser.test.js', () => {
   let app;
   let csrf;
   let cookies;
